# Refine estimates for estimateCardinality for TableReference