@jianqiao 2016-11-07
LGTM. Merging.